// src/modules/Permissions/enums.ts
export type BipPermissionKey =
  | StudioPermissionEnum
  | CollectionPermissionEnum
  | CanvasPermissionEnum;

// Studio Permissions
export enum StudioPermissionEnum {
  STUDIO_CREATE_DELETE_ROLE = "STUDIO_CREATE_DELETE_ROLE",
  STUDIO_EDIT_STUDIO_PROFILE = "STUDIO_EDIT_STUDIO_PROFILE",
  STUDIO_ADD_REMOVE_USER_TO_ROLE = "STUDIO_ADD_REMOVE_USER_TO_ROLE",
  STUDIO_MANAGE_INTEGRATION = "STUDIO_MANAGE_INTEGRATION",
  STUDIO_CREATE_COLLECTION = "STUDIO_CREATE_COLLECTION",
  STUDIO_METADATA_UPDATE = "STUDIO_METADATA_UPDATE",
  STUDIO_CHANGE_CANVAS_COLLECTION_POSITION = "STUDIO_CHANGE_CANVAS_COLLECTION_POSITION",
  CAN_MANAGE_BILLING = "CAN_MANAGE_BILLING",
}

// Collection Permisions
export enum CollectionPermissionEnum {
  COLLECTION_MEMBERSHIP_MANAGE = "COLLECTION_MEMBERSHIP_MANAGE",
  COLLECTION_PUBLIC_ACCESS_CHANGE = "COLLECTION_PUBLIC_ACCESS_CHANGE",
  COLLECTION_MANAGE_PERMS = "COLLECTION_MANAGE_PERMS",
  COLLECTION_DELETE = "COLLECTION_DELETE",
  COLLECTION_OVERRIDE_STUDIO_MODE_ROLE = "COLLECTION_OVERRIDE_STUDIO_MODE_ROLE",
  COLLECTION_VIEW_METADATA = "COLLECTION_VIEW_METADATA",
  COLLECTION_EDIT_NAME = "COLLECTION_EDIT_NAME",
}

// Canvas Permission
export enum CanvasPermissionEnum {
  CANVAS_BRANCH_VIEW = "CANVAS_BRANCH_VIEW",
  CANVAS_BRANCH_EDIT = "CANVAS_BRANCH_EDIT",
  CANVAS_BRANCH_DELETE = "CANVAS_BRANCH_DELETE",
  CANVAS_BRANCH_ADD_COMMENT = "CANVAS_BRANCH_ADD_COMMENT",
  CANVAS_BRANCH_CHANGE_PUBLIC_ACCESS = "CANVAS_BRANCH_CHANGE_PUBLIC_ACCESS",
  CANVAS_BRANCH_EDIT_NAME = "CANVAS_BRANCH_EDIT_NAME",
  CANVAS_BRANCH_ADD_REACTION = "CANVAS_BRANCH_ADD_REACTION",
  CANVAS_BRANCH_CREATE_REEL = "CANVAS_BRANCH_CREATE_REEL",
  CANVAS_BRANCH_COMMENT_ON_REEL = "CANVAS_BRANCH_COMMENT_ON_REEL",
  CANVAS_BRANCH_REACT_TO_REEL = "CANVAS_BRANCH_REACT_TO_REEL",
  CANVAS_BRANCH_MANAGE_PERMS = "CANVAS_BRANCH_MANAGE_PERMS",
  CANVAS_BRANCH_MANAGE_MERGE_REQUESTS = "CANVAS_BRANCH_MANAGE_MERGE_REQUESTS",
  CANVAS_BRANCH_MANAGE_PUBLISH_REQUESTS = "CANVAS_BRANCH_MANAGE_PUBLISH_REQUESTS",
  CANVAS_BRANCH_CREATE_MERGE_REQUEST = "CANVAS_BRANCH_CREATE_MERGE_REQUEST",
  CANVAS_BRANCH_CREATE_PUBLISH_REQUEST = "CANVAS_BRANCH_CREATE_PUBLISH_REQUEST",
  CANVAS_BRANCH_CLONE = "CANVAS_BRANCH_CLONE",
  CANVAS_BRANCH_RESOLVE_COMMENTS = "CANVAS_BRANCH_RESOLVE_COMMENTS",
  CANVAS_BRANCH_MANAGE_CONTENT = "CANVAS_BRANCH_MANAGE_CONTENT",
}

// Canvas Permission Groups
export enum CanvasPermissionGroupEnum {
  NONE = "pg_canvas_branch_none",
  VIEW = "pg_canvas_branch_view",
  COMMENT = "pg_canvas_branch_comment",
  EDIT = "pg_canvas_branch_edit",
  MODERATE = "pg_canvas_branch_moderate",
  VIEW_METADATA = "pg_canvas_branch_view_metadata",
}

export enum CollectionPermissionGroupEnum {
  NONE = "pg_collection_none",
  VIEW = "pg_collection_view",
  COMMENT = "pg_collection_comment",
  EDIT = "pg_collection_edit",
  MODERATE = "pg_collection_moderate",
  VIEW_METADATA = "pg_collection_view_metadata",
}

export enum PermissionTreeContextEnum {
  Member = "member",
  Role = "role",
}